+++ CHANGELOG.md
@@ -12,7 +12,9 @@
 * Generating xtream stream urls from m3u input.
 * Reverse proxy implementation for m3u playlist.
 * Mapper can now set `epg_channel_id`.
-* Added environment variables for User Credentials `username`, `password` and `token` in format `${env:<EnvVarName>}` where `<EnvVarName>` should be replaced. 
+* Added environment variables for User Credentials `username`, `password` and `token` in format `${env:<EnvVarName>}` where `<EnvVarName>` should be replaced.
+* Added `web_ui_enabled` to `config.yml`. Default is `true`. Set to `false` to disable webui.
+* Added `web_auth_enabled` to `config.yml`. Default is `false`. Set to `true` to enable webui authentication.
 
 # v1.1.8(2024-03-06)
 * Fixed WebUI Option-Select